[nnn] [system-font-v1] use system's default font in more places
This changes the font when the "Use system's default font" setting is
enabled by the user. This patch should not change anything otherwise.

Use system-ui/sans-serif/monospace font family and do not try to
hardcode any font names.

The option is also currently not honored in <code> elements. Fix it.