[use-remote-links-v19] Use original account/status URLs in DOM
Unbreak "Open Link in New Tab" on remote account or status links.

Whenever possible, let <a> href attributes point at the original "url"
of the remote account or status instead of the local instance. This
patch does not change the onclick event handler, so left-clicking on it
will still open the content within the JavaScript application.