[better-search-v1] Use N-gram tokenizer for status texts
The standard tokenizer and the uax_url_email tokenizer, which is based
on the standard tokenizer, don't know how to tokenize Japanese text.

I think using Kuromoji or Sudachi might be a better idea, but let's see
if the builtin N-gram tokenizer works, which doesn't require me to
install any additional dependencies.